STM32CubeExpansion_Cloud_AZURE_V2.3.0/Projects/B-U585I-IOT02A/Applications/TFM_Azure_IoT/AzureScripts/requirements.txt
pathlib==1.0.1
pyserial
azure-cli
halo